// src/app/model/ItemsModel.ts
export class ItemsModel {
    ItemId: number;
    OrganizationId: number;
    OrganizationName: string;
    ItemCode: string;
    ItemDescr: string;
    LCategoryId: number;
    LSubCategoryId: number;
    LUnitId: number;
    CategoryName: string;
    SubCategoryName: string;
    Unit: string;
    UnitPrice: number;
    Note: string;
    IsActive: boolean;
    CreatedDate?: Date;

}